WILSON 3.1 actual in progress

+++ 1.1.py
@@ -10,11 +10,11 @@ else:
 
     api.single_fly_takeoff()
     api.single_fly_Qrcode_align(0, 0)
-    api.single_fly_down(20/CM)
-    api.single_fly_forward(200 / CM) #subject to changes
-    api.single_fly_up(200 /CM)
-    api.single_fly_forward(120 / CM)
-    api.single_fly_down(150/ CM)
+    api.single_fly_down(int(20/CM))
+    api.single_fly_forward(int(200 / CM)) #subject to changes
+    api.single_fly_up(int(200 /CM))
+    api.single_fly_forward(int(120 / CM))
+    api.single_fly_down(int(150/ CM))
     # time.sleep(2)
-    api.single_fly_forward(200/ CM)
+    api.single_fly_forward(int(200/ CM))
     api.single_fly_touchdown()
